+++ web-wallet/src/routes/(welcome)/login/__tests__/page.spec.js
@@ -119,10 +119,8 @@ describe("Login", async () => {
 			expect(gotoSpy).toHaveBeenCalledWith("/dashboard");
 		});
 
-		it("should trigger a notice if the user inputs a valid mnemonic different from the last one used clear local data and direct to dashboard if accepted", async () => {
-			const newUserId = generateMnemonic();
-
-			settingsStore.update(setKey("userId", newUserId));
+		it("should trigger a notice if the user inputs a valid mnemonic different from the last one used, then clear local data and redirect to dashboard if the notice is accepted", async () => {
+			settingsStore.update(setKey("userId", "some-user-id"));
 			confirmSpy.mockReturnValue(true);
 
 			const { container } = render(Login, {});
@@ -145,14 +143,8 @@ describe("Login", async () => {
 			expect(gotoSpy).toHaveBeenCalledWith("/dashboard");
 		});
 
-		it("should trigger a notice if the user inputs a valid mnemonic different from the last one used and show an error if declined", async () => {
-			const newUserId = generateMnemonic();
-
-			settingsStore.update(setKey("userId", newUserId));
-
-			const errorMessage = "Existing wallet detected";
-
-			walletResetSpy.mockRejectedValueOnce(new Error(errorMessage));
+		it("should trigger a notice if the user inputs a valid mnemonic different from the last one used and show an error if the notice is declined", async () => {
+			settingsStore.update(setKey("userId", "some-user-id"));
 			confirmSpy.mockReturnValue(false);
 
 			const { container } = render(Login, {});
@@ -180,12 +172,12 @@ describe("Login", async () => {
 			expect(get(settingsStore).userId).not.toBe(userId);
 			expect(initSpy).not.toHaveBeenCalled();
 			expect(gotoSpy).not.toHaveBeenCalled();
-			expect(errorElement?.textContent).toBe(errorMessage);
+			expect(errorElement?.textContent).toBe("Existing wallet detected");
 			expect(textInput).toHaveFocus();
 			expect(selectedText).toBe(textInput.value);
 		});
 
-		it("should not clear local data if the entered mnemonic is the last one used", async () => {
+		it("should not show a notice and should not clear local data if the entered mnemonic is the last one used", async () => {
 			settingsStore.update(setKey("userId", userId));
 
 			const { container } = render(Login, {});
@@ -206,6 +198,41 @@ describe("Login", async () => {
 			expect(initSpy).toHaveBeenCalledWith(expect.any(Wallet));
 			expect(gotoSpy).toHaveBeenCalledTimes(1);
 			expect(gotoSpy).toHaveBeenCalledWith("/dashboard");
+		});
+
+		it("should show an error if the clearing of local data fails", async () => {
+			settingsStore.update(setKey("userId", ""));
+
+			const errorMessage = "Failed to delete data";
+
+			walletResetSpy.mockRejectedValueOnce(new Error(errorMessage));
+
+			const { container } = render(Login, {});
+			const form = getAsHTMLElement(container, "form");
+			const textInput = getTextInput(container);
+
+			expect(getErrorElement()).toBeNull();
+
+			await fireEvent.input(textInput, { target: { value: mnemonic } });
+			await fireEvent.submit(form, { currentTarget: form });
+
+			const errorElement = await vi.waitUntil(getErrorElement);
+			const selectedText = textInput.value.substring(
+				Number(textInput.selectionStart),
+				Number(textInput.selectionEnd)
+			);
+
+			expect(confirmSpy).not.toHaveBeenCalled();
+			expect(getWalletSpy).toHaveBeenCalledTimes(1);
+			expect(getWalletSpy).toHaveBeenCalledWith(seed);
+			expect(walletResetSpy).toHaveBeenCalledTimes(1);
+			expect(settingsResetSpy).not.toHaveBeenCalled();
+			expect(get(settingsStore).userId).not.toBe(userId);
+			expect(initSpy).not.toHaveBeenCalled();
+			expect(gotoSpy).not.toHaveBeenCalled();
+			expect(errorElement?.textContent).toBe(errorMessage);
+			expect(textInput).toHaveFocus();
+			expect(selectedText).toBe(textInput.value);
 		});
 
 		it("should trim the entered mnemonic before validating it", async () => {
@@ -270,7 +297,107 @@ describe("Login", async () => {
 			expect(selectedText).toBe(textInput.value);
 		});
 
-		it("should not clear local data if the entered password is for the last used mnemonic", async () => {
+		/**
+		 * This is not a possible situation, in theory, but
+		 * the workflow is able to deal with it.
+		 */
+		it("should clear local data and redirect to the dashboard if the user inputs the correct password with no prior wallet created", async () => {
+			settingsStore.update(setKey("userId", ""));
+
+			const { container } = render(Login, {});
+			const form = getAsHTMLElement(container, "form");
+			const textInput = getTextInput(container);
+
+			await fireEvent.input(textInput, { target: { value: pwd } });
+			await fireEvent.submit(form, { currentTarget: form });
+
+			expect(getErrorElement()).toBeNull();
+
+			await vi.waitUntil(() => gotoSpy.mock.calls.length > 0);
+
+			expect(getErrorElement()).toBeNull();
+			expect(confirmSpy).not.toHaveBeenCalled();
+			expect(getWalletSpy).toHaveBeenCalledTimes(1);
+			expect(getWalletSpy).toHaveBeenCalledWith(seed);
+			expect(walletResetSpy).toHaveBeenCalledTimes(1);
+			expect(settingsResetSpy).toHaveBeenCalledTimes(1);
+			expect(get(settingsStore).userId).toBe(userId);
+			expect(initSpy).toHaveBeenCalledTimes(1);
+			expect(initSpy).toHaveBeenCalledWith(expect.any(Wallet));
+			expect(gotoSpy).toHaveBeenCalledTimes(1);
+			expect(gotoSpy).toHaveBeenCalledWith("/dashboard");
+		});
+
+		/**
+		 * This is not a possible situation, in theory, but
+		 * the workflow is able to deal with it.
+		 */
+		it("should ask to clear local data and redirect to the dashboard if the user inputs the correct password for a mnemonic different from the last one used", async () => {
+			settingsStore.update(setKey("userId", "some-fake-id"));
+			confirmSpy.mockReturnValue(true);
+
+			const { container } = render(Login, {});
+			const form = getAsHTMLElement(container, "form");
+			const textInput = getTextInput(container);
+
+			await fireEvent.input(textInput, { target: { value: pwd } });
+			await fireEvent.submit(form, { currentTarget: form });
+
+			expect(getErrorElement()).toBeNull();
+
+			await vi.waitUntil(() => gotoSpy.mock.calls.length > 0);
+
+			expect(confirmSpy).toHaveBeenCalledOnce();
+			expect(getWalletSpy).toHaveBeenCalledTimes(1);
+			expect(getWalletSpy).toHaveBeenCalledWith(seed);
+			expect(walletResetSpy).toHaveBeenCalledTimes(1);
+			expect(settingsResetSpy).toHaveBeenCalledTimes(1);
+			expect(get(settingsStore).userId).toBe(userId);
+			expect(initSpy).toHaveBeenCalledTimes(1);
+			expect(initSpy).toHaveBeenCalledWith(expect.any(Wallet));
+			expect(gotoSpy).toHaveBeenCalledTimes(1);
+			expect(gotoSpy).toHaveBeenCalledWith("/dashboard");
+		});
+
+		/**
+		 * This is not a possible situation, in theory, but
+		 * the workflow is able to deal with it.
+		 */
+		it("should ask to clear local data and show an error if the user inputs the correct password for a mnemonic different from the last one used and decline the notice", async () => {
+			settingsStore.update(setKey("userId", "some-fake-id"));
+			confirmSpy.mockReturnValue(false);
+
+			const { container } = render(Login, {});
+			const form = getAsHTMLElement(container, "form");
+			const textInput = getTextInput(container);
+
+			expect(getErrorElement()).toBeNull();
+
+			await fireEvent.input(textInput, { target: { value: pwd } });
+			await fireEvent.submit(form, { currentTarget: form });
+			await vi.waitUntil(() => confirmSpy.mock.calls.length > 0);
+
+			expect(confirmSpy).toHaveBeenCalledOnce();
+
+			const errorElement = await vi.waitUntil(getErrorElement);
+			const selectedText = textInput.value.substring(
+				Number(textInput.selectionStart),
+				Number(textInput.selectionEnd)
+			);
+
+			expect(getWalletSpy).toHaveBeenCalledTimes(1);
+			expect(getWalletSpy).toHaveBeenCalledWith(seed);
+			expect(walletResetSpy).not.toHaveBeenCalled();
+			expect(settingsResetSpy).not.toHaveBeenCalled();
+			expect(get(settingsStore).userId).not.toBe(userId);
+			expect(initSpy).not.toHaveBeenCalled();
+			expect(gotoSpy).not.toHaveBeenCalled();
+			expect(errorElement?.textContent).toBe("Existing wallet detected");
+			expect(textInput).toHaveFocus();
+			expect(selectedText).toBe(textInput.value);
+		});
+
+		it("should not show a notice and clear local data if the entered password is for the last used mnemonic", async () => {
 			settingsStore.update(setKey("userId", userId));
 
 			const { container } = render(Login, {});
@@ -291,6 +418,42 @@ describe("Login", async () => {
 			expect(initSpy).toHaveBeenCalledWith(expect.any(Wallet));
 			expect(gotoSpy).toHaveBeenCalledTimes(1);
 			expect(gotoSpy).toHaveBeenCalledWith("/dashboard");
+		});
+
+		it("should show an error if the clearing of local data fails", async () => {
+			settingsStore.update(setKey("userId", "some-fake-id"));
+			confirmSpy.mockReturnValue(true);
+
+			const errorMessage = "Failed to delete data";
+
+			walletResetSpy.mockRejectedValueOnce(new Error(errorMessage));
+
+			const { container } = render(Login, {});
+			const form = getAsHTMLElement(container, "form");
+			const textInput = getTextInput(container);
+
+			expect(getErrorElement()).toBeNull();
+
+			await fireEvent.input(textInput, { target: { value: pwd } });
+			await fireEvent.submit(form, { currentTarget: form });
+			await vi.waitUntil(() => confirmSpy.mock.calls.length > 0);
+
+			const errorElement = await vi.waitUntil(getErrorElement);
+			const selectedText = textInput.value.substring(
+				Number(textInput.selectionStart),
+				Number(textInput.selectionEnd)
+			);
+
+			expect(getWalletSpy).toHaveBeenCalledTimes(1);
+			expect(getWalletSpy).toHaveBeenCalledWith(seed);
+			expect(walletResetSpy).toHaveBeenCalledTimes(1);
+			expect(settingsResetSpy).not.toHaveBeenCalled();
+			expect(get(settingsStore).userId).not.toBe(userId);
+			expect(initSpy).not.toHaveBeenCalled();
+			expect(gotoSpy).not.toHaveBeenCalled();
+			expect(errorElement?.textContent).toBe(errorMessage);
+			expect(textInput).toHaveFocus();
+			expect(selectedText).toBe(textInput.value);
 		});
 
 		it("should trim the entered password before validating it", async () => {